+++ NEWS.md
@@ -1,3 +1,6 @@
+# explor 0.3.7
+
+
 # explor 0.3.6
 
 * Add support for textmodel_ca